+++ GraphQL/Error/FormInvalidException.php
@@ -1,14 +1,22 @@
 <?php
 
+/*
+ * This file is part of EC-CUBE
+ *
+ * Copyright(c) EC-CUBE CO.,LTD. All Rights Reserved.
+ *
+ * http://www.ec-cube.co.jp/
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
 
 namespace Plugin\Api\GraphQL\Error;
-
 
 use GraphQL\Error\ClientAware;
 
 class FormInvalidException extends \Exception implements ClientAware
 {
-
     public function isClientSafe()
     {
         return true;
