+++ Resource/template/admin/OAuth/authorization.twig
